it worked!!
The issue (at least from this test) was creating the
instances in different subsets. That is OK if we can use
DNS from the router, but not if we are using the internal
addresses (as userentes is). The fix was still to create
3 subnets / AZones (it will get angry if you do not) but
to only specify one for the autoscaling group! Could it
have been luck? Maybe - but 7 working seems like it is not

Signed-off-by: vsoch <vsoch@users.noreply.github.com>